requirements.txt
# Core dependencies
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.1.1
requests==2.32.5
setuptools==80.9.0
urllib3==2.5.0
wheel==0.45.1

# Reddit data collection
praw==7.8.1
prawcore==2.4.0
update-checker==0.18.0
websocket-client==1.9.0
schedule==1.2.2

# Sentiment analysis
vaderSentiment==3.3.2

# Database (Supabase)
supabase==2.3.0
postgrest==0.13.2
realtime==1.0.0
storage3==0.7.0
gotrue==1.3.0
supafunc==0.3.1

# Vector embeddings
sentence-transformers==3.3.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
transformers>=4.35.0
huggingface-hub>=0.19.0

# LLM integration
groq==0.4.0

# Chat interface
streamlit>=1.29.0
streamlit-chat>=0.1.1

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
